Split Drag and Drop components into separate files
